fix templates
