+++ modules/templates/templates.go
@@ -4,5 +4,5 @@
 
 package templates
 
-//go:generate go run ../embed/main.go templates
+//go:generate go run main.go
 //go:generate go fmt bindata.go
